Module Sidekick_abstract_solver.Check_res

Result of solving for the current set of clauses

type value = Sidekick_core.Term.t

type sat_result = {

get_value : Sidekick_core.Term.t -> value option;	(*
Value for this term
*)

iter_classes : (Sidekick_core.Term.t Iter.t * value) Iter.t;	(*
All equivalence classes in the congruence closure
*)

eval_lit : Sidekick_core.Lit.t -> bool option;	(*
Evaluate literal
*)

iter_true_lits : Sidekick_core.Lit.t Iter.t;	(*
Iterate on literals that are true in the trail
*)

}

Satisfiable

type unsat_result = {

unsat_core : unit -> Sidekick_core.Lit.t Iter.t;	(*
Unsat core (subset of assumptions), or empty
*)

unsat_proof : unit -> Sidekick_proof.Step.id option;	(*
Proof step for the empty clause
*)

}

Unsatisfiable

type t =

| Sat of sat_result

| Unsat of unsat_result

| Unknown of Unknown.t	(*
Unknown, obtained after a timeout, memory limit, etc.
*)

Result of calling "check"

val pp : Sidekick_core.Fmt.t -> t -> unit