Module type Linear_expr.S

module C : Sidekick_arith_lra__.Linear_expr_intf.COEFF

Coeficients used. Can be integers as well as rationals.

module Var : Sidekick_arith_lra__.Linear_expr_intf.VAR

Variables used in expressions.

type var = Var.t

The type of variables appearing in expressions.

module Var_map : CCMap.S with type Var_map.key = var

Maps from variables, used for expressions as well as substitutions.

type subst = C.t Var_map.t

Type for substitutions.

module Comb : sig ... end

Combinations.

Linear expressions.

module Expr : sig ... end

Linear expressions represent linear arithmetic expressions as a linear combination and a constant.

module Constr : sig ... end