Parameter Make.M

type t

Some type with a monoid structure

include Sidekick_sigs.PRINT with type t := t

val pp : t Sidekick_sigs.printer

type state

val create : CC.t -> state

Initialize state from the congruence closure

val name : string

name of the monoid structure (short)

val of_term : CC.t -> state -> E_node.t -> Sidekick_core.Term.t -> t option * (E_node.t * t) list

of_term n t, where t is the Term.t annotating node n, must return maybe_m, l, where:

maybe_m = Some m if t has monoid value m; otherwise maybe_m=None

l is a list of (u, m_u) where each u's Term.t is a direct subterm of t and m_u is the monoid value attached to u.

val merge : CC.t -> state -> E_node.t -> t -> E_node.t -> t -> Expl.t -> (t * CC.Handler_action.t list, CC.Handler_action.conflict) Stdlib.result

Monoidal combination of two values.

merge cc n1 mon1 n2 mon2 expl returns the result of merging monoid values mon1 (for class n1) and mon2 (for class n2) when n1 and n2 are merged with explanation expl.

returns
Ok mon if the merge is acceptable, annotating the class of n1 ∪ n2; or Error expl' if the merge is unsatisfiable. expl' can then be used to trigger a conflict and undo the merge.