Module Sidekick_cc.Result_action

Result Actions.

Actions returned by the congruence closure after calling check.

type t =

| Act_propagate of {

lit : Sidekick_core.Lit.t;

reason : propagation_reason;

}	(*
propagate (Lit.t, reason) declares that reason() => Lit.t is a tautology.

reason() should return a list of literals that are currently true, as well as a proof.

Lit.t should be a literal of interest (see S.set_as_lit).

This function might never be called, a congruence closure has the right to not propagate and only trigger conflicts.
*)

type conflict =

| Conflict of Sidekick_core.Lit.t list * Sidekick_core.Proof_term.step_id	(*
raise_conflict (c,pr) declares that c is a tautology of the theory of congruence.

parameter pr
the proof of c being a tautology

*)

type or_conflict = ( t list, conflict ) Stdlib.result