Module Sidekick_core

Main Signatures.

Theories and concrete solvers rely on an environment that defines several important types:

types

terms (to represent logic expressions and formulas)

a congruence closure instance

a bridge to some SAT solver

In this module we define most of the main signatures used throughout Sidekick.

module Fmt = CCFormat

Re-exports from core-logic

module Const = Sidekick_core_logic.Const

module Term : sig ... end

module Gensym : sig ... end

Fresh symbol generation

view

module Bool_view : sig ... end

Boolean-oriented view of terms

module CC_view : sig ... end

View terms through the lens of a Congruence Closure

module Default_cc_view : sig ... end

Main modules

module Bvar = Sidekick_core_logic.Bvar

module Lit : sig ... end

Literals

module Proof_step : sig ... end

module Proof_core : sig ... end

Core proofs for SMT and congruence closure.

module Proof_sat : sig ... end

SAT-solver proof emission.

module Proof_trace : sig ... end

Proof traces.

module Proof_term : sig ... end

Proof terms.

module Subst = Sidekick_core_logic.Subst

module Var = Sidekick_core_logic.Var

module Box : sig ... end

module Clause_tracer : sig ... end

Tracer for clauses and literals

exception Resource_exhausted