Module Sidekick_core.Gensym

Fresh symbol generation

type term = Sidekick_core_logic.Term.t

type ty = Sidekick_core_logic.Term.t

type t

Fresh symbol generator.

The theory needs to be able to create new terms with fresh names, to be used as placeholders for complex formulas during Tseitin encoding.

val create : Sidekick_core_logic.Term.store -> t

New (stateful) generator instance.

val const_decoders : Sidekick_core_logic.Const.decoders

val fresh_term : t -> pre:string -> ty -> term

Make a fresh term of the given type

val reset : t -> unit

Reset to initial state