Module SI.Simplify

Simplify terms

type t

val tst : t -> term_store

val ty_st : t -> ty_store

val clear : t -> unit

Reset internal cache, etc.

val with_proof : t -> (proof -> unit) -> unit

type hook = t -> term -> term option

Given a term, try to simplify it. Return None if it didn't change.

A simple example could be a hook that takes a term t, and if t is app "+" (const x) (const y) where x and y are number, returns Some (const (x+y)), and None otherwise.

val normalize : t -> term -> term option

Normalize a term using all the hooks. This performs a fixpoint, i.e. it only stops when no hook applies anywhere inside the term.

val normalize_t : t -> term -> term

Normalize a term using all the hooks, along with a proof that the simplification is correct. returns t, refl t if no simplification occurred.