Module Sidekick_core.Proof_trace

Proof traces.

A proof trace is a log of all the deductive reasoning steps made by the SMT solver and other reasoning components. It essentially stores a DAG of all these steps, where each step points (via step_id) to its premises.

type lit = Lit.t

type step_id = Proof_step.id

Identifier for a tracing step (like a unique ID for a clause previously added/proved)

module Step_vec : Sidekick_util.Vec_sig.BASE with type elt = step_id

A vector indexed by steps.

Traces

type t

The proof trace itself.

A proof trace is a log of all deductive steps taken by the solver, so we can later reconstruct a certificate for proof-checking.

Each step in the proof trace should be a valid lemma (of its theory) or a valid consequence of previous steps.

val enabled : t -> bool

Is proof tracing enabled?

val add_step : t -> Proof_term.delayed -> step_id

Create a new step in the trace.

val add_unsat : t -> step_id -> unit

Signal "unsat" result at the given proof

val delete : t -> step_id -> unit

Forget a step that won't be used in the rest of the trace. Only useful for performance/memory considerations.

val close : t -> unit

close p closes the proof, and can dispose of underlying resources

Dummy backend

val dummy_step_id : step_id

val dummy : t

Dummy proof trace, logs nothing.

Dynamic interface

module type DYN = sig ... end

val make : (module DYN) -> t