Module Term.Trace_reader

module Tr = Sidekick_trace

module Dec = Sidekick_util.Ser_decode

type term_ref = Tr.entry_id

type const_decoders = Sidekick_core_logic.Const.decoders

type t

val create : ?const_decoders:const_decoders list -> source:Tr.Source.t -> Sidekick_core_logic.Term.store -> t

val store : t -> Sidekick_core_logic.Term.store

val add_const_decoders : t -> const_decoders -> unit

val read_term : t -> term_ref -> (Sidekick_core_logic.Term.t, string) Stdlib.result

val read_term_err : t -> term_ref -> (Sidekick_core_logic.Term.t, Sidekick_util.Ser_decode.Error.t) Stdlib.result

val read_term_exn : t -> term_ref -> Sidekick_core_logic.Term.t

raises Error.Error
if it fails

val deref_term : t -> Sidekick_core_logic.Term.t -> Sidekick_core_logic.Term.t

deref_term reader t dereferences the root node of t. If t is Ref id, this returns the term at id instead.

raises Error.Error
if it fails