Module Term.Tracer

Emit terms in traces.

Traces will contains terms, encoded as a DAG. Each subterm is its own event, and gets a term identifier used in other subsequent entries to refer to it.

module Tr = Sidekick_trace

type term_ref = Tr.entry_id

type Tr.entry_view +=

| T_ty of int

| T_app of term_ref * term_ref

| T_var of string * term_ref

| T_bvar of int * term_ref

| T_const of {

c : Sidekick_core_logic.Const.view;

c_ops : Sidekick_core_logic.Const.Ops.t;

ty : term_ref;

}

| T_lam of {

v_name : string;

v_ty : term_ref;

body : term_ref;

}

| T_pi of {

v_name : string;

v_ty : term_ref;

body : term_ref;

}

class type t = object ... end

class dummy : object ... end

Dummy implementation, returns Tr.Entry_id.dummy

class concrete : sink:Tr.Sink.t -> t

Concrete implementation of t

val create : sink:Tr.Sink.t -> unit -> t

create ~sink () makes a tracer that will write terms into the given sink.

val emit : t -> Sidekick_core_logic.Term.t -> term_ref

val emit' : t -> Sidekick_core_logic.Term.t -> unit