Module type Sidekick_core.MONOID_ARG

module SI : SOLVER_INTERNAL

type t

val pp : t Fmt.printer

val name : string

name of the monoid's value (short)

val of_term : SI.CC.t -> SI.CC.N.t -> SI.T.Term.t -> t option * (SI.CC.N.t * t) list

of_term n t, where t is the term annotating node n, returns maybe_m, l, where:

maybe_m = Some m if t has monoid value m; otherwise maybe_m=None

l is a list of (u, m_u) where each u's term is a direct subterm of t and m_u is the monoid value attached to u.

val merge : SI.CC.t -> SI.CC.N.t -> t -> SI.CC.N.t -> t -> SI.CC.Expl.t -> (t, SI.CC.Expl.t) Stdlib.result