Module type Solver_internal.PREPROCESS_ACTS

val proof : proof

val mk_lit_nopreproc : ?sign:bool -> term -> lit

mk_lit t creates a new literal for a boolean term t.

val mk_lit : ?sign:bool -> term -> lit * proof_step option

mk_lit t creates a new literal for a boolean term t. Also returns an optional proof of preprocessing, which if present is the proof of |- t = lit with lit the result.

val add_clause : lit list -> proof_step -> unit

pushes a new clause into the SAT solver.

val add_lit : ?default_pol:bool -> lit -> unit

Ensure the literal will be decided/handled by the SAT solver.