Module Sidekick_core_logic.T_builtins

Core builtins

type Sidekick_core_logic__.Types_.const_view +=

| C_bool

| C_eq

| C_ite

| C_not

| C_true

| C_false

| C_proof

val bool : Term.store -> Term.t

val proof : Term.store -> Term.t

val c_not : Term.store -> Term.t

val c_eq : Term.store -> Term.t

val c_ite : Term.store -> Term.t

val true_ : Term.store -> Term.t

val false_ : Term.store -> Term.t

val bool_val : Term.store -> bool -> Term.t

val const_decoders : Const.decoders

val eq : Term.store -> Term.t -> Term.t -> Term.t

eq a b is a = b

val not : Term.store -> Term.t -> Term.t

val ite : Term.store -> Term.t -> Term.t -> Term.t -> Term.t

ite a b c is if a then b else c

val is_eq : Term.t -> bool

val is_bool : Term.t -> bool

val abs : Term.store -> Term.t -> bool * Term.t

abs t returns an "absolute value" for the term, along with the sign of t.

The idea is that we want to turn not a into (false, a), or (a != b) into (false, a=b). For terms without a negation this should return (true, t).

val as_bool_val : Term.t -> bool option

val open_eq : Term.t -> (Term.t * Term.t) option

open_eq (a=b) returns Some (a,b), None for other terms.