Module Sidekick_model

Models

A model can be produced when the solver is found to be in a satisfiable state after a call to solve.

type t

val empty : t

val is_empty : t -> bool

val mem : t -> Sidekick_core.Term.t -> bool

val find : t -> Sidekick_core.Term.t -> Sidekick_core.Term.t option

val eval : t -> Sidekick_core.Term.t -> Sidekick_core.Term.t option

val add : Sidekick_core.Term.t -> Sidekick_core.Term.t -> t -> t

val keys : t -> Sidekick_core.Term.t Iter.t

val to_iter : t -> (Sidekick_core.Term.t * Sidekick_core.Term.t) Iter.t

include Sidekick_sigs.PRINT with type t := t

val pp : t Sidekick_sigs.printer