Module Sidekick_quip.Proof

type id = int

A reference to a previously defined object in the proof

module Ty : sig ... end

Representation of types

module Fun : sig ... end

module Cstor = Fun

module T : sig ... end

Representation of terms, with explicit sharing

type term = T.t

type ty = Ty.t

module Lit : sig ... end

type clause = Lit.t list

type t =

| Unspecified

| Sorry

| Sorry_c of clause

| Named of string

| Refl of term

| CC_lemma_imply of t list * term * term

| CC_lemma of clause

| Assertion of term

| Assertion_c of clause

| Hres of t * hres_step list

| Res of term * t * t

| Res1 of t * t

| Paramod1 of t * t

| Rup of clause * t list

| Clause_rw of {

res : clause;

c0 : t;

using : t list;
(*
the rewriting equations/atoms
*)

}

| DT_isa_split of ty * term list

| DT_isa_disj of ty * term * term

| DT_cstor_inj of Cstor.t * int * term list * term list

| Bool_true_is_true

| Bool_true_neq_false

| Bool_eq of term * term

| Bool_c of bool_c_name * term list

| Ite_true of term

| Ite_false of term

| LRA of clause

| Composite of {

assumptions : (string * Lit.t) list;

steps : composite_step array;

}

and bool_c_name = string

and composite_step =

| S_step_c of {

name : string;

res : clause;

proof : t;

}
(*
A named step in Composite, with the expected result. This decouples the checking of the sub-proof, from its use in the rest of the proof, as we can use res even if checking proof failed.
*)

| S_step_anon of {

name : string;

proof : t;

}
(*
A named intermediate proof, to be reused in subsequent proofs. Unlike S_step_c we do not specify the expected result so if this fails, any proof downstream will also fail.
*)

| S_define_t of term * term

| S_define_t_name of string * term

and hres_step =

| R of {

pivot : term;

p : t;

}

| R1 of t

| P of {

lhs : term;

rhs : term;

p : t;

}

| P1 of t

val r : t -> pivot:term -> hres_step

val r1 : t -> hres_step

val p : t -> lhs:term -> rhs:term -> hres_step

val p1 : t -> hres_step

val stepc : name:string -> clause -> t -> composite_step

val step_anon : name:string -> t -> composite_step

val deft : term -> term -> composite_step

val deft_name : string -> term -> composite_step

val is_trivial_refl : t -> bool

val default : t

val sorry_c : Lit.t Iter.t -> t

val sorry_c_l : clause -> t

val sorry : t

val refl : term -> t

val ref_by_name : string -> t

val cc_lemma : clause -> t

val cc_imply_l : t list -> term -> term -> t

val cc_imply2 : t -> t -> term -> term -> t

val assertion : term -> t

val assertion_c : Lit.t Iter.t -> t

val assertion_c_l : clause -> t

val rup : clause -> t list -> t

val clause_rw : t -> res:clause -> using:t list -> t

val composite_a : ?assms:(string * Lit.t) list -> composite_step array -> t

val composite_l : ?assms:(string * Lit.t) list -> composite_step list -> t

val composite_iter : ?assms:(string * Lit.t) list -> composite_step Iter.t -> t

val isa_split : ty -> term Iter.t -> t

val isa_disj : ty -> term -> term -> t

val cstor_inj : Cstor.t -> int -> term list -> term list -> t

val ite_true : term -> t

val ite_false : term -> t

val true_is_true : t

val true_neq_false : t

val bool_eq : term -> term -> t

val bool_c : bool_c_name -> term list -> t

val hres_l : t -> hres_step list -> t

val hres_iter : t -> hres_step Iter.t -> t

val res : pivot:term -> t -> t -> t

val res1 : t -> t -> t

val paramod1 : t -> t -> t

val lra_l : clause -> t

val lra : Lit.t Iter.t -> t