Module type Solver_intf.SAT_STATE

type lit

Literals (signed boolean atoms)

val eval : lit -> bool

Returns the valuation of a lit in the current state of the sat solver.

raises UndecidedLit
if the literal is not decided

val eval_level : lit -> bool * int

Return the current assignement of the literals, as well as its decision level. If the level is 0, then it is necessary for the literal to have this value; otherwise it is due to choices that can potentially be backtracked.

raises UndecidedLit
if the literal is not decided

val iter_trail : (lit -> unit) -> unit

Iter through the lits in order of decision/propagation (starting from the first propagation, to the last propagation).