Module Sidekick_sat__.Solver_intf

type 'a printer = Stdlib.Format.formatter -> 'a -> unit

module type SAT_STATE = sig ... end

type 'form sat_state = (module SAT_STATE with type lit = 'form)

The type of values returned when the solver reaches a SAT state.

module type UNSAT_STATE = sig ... end

type ('lit, 'clause) unsat_state = (module UNSAT_STATE with type clause = 'clause and type lit = 'lit)

The type of values returned when the solver reaches an UNSAT state.

type same_sign = bool

This type is used during the normalisation of lits. true means the literal stayed the same, false that its sign was flipped.

type ('lit, 'proof) reason =

| Consequence of unit -> 'lit list * 'proof

The type of reasons for propagations of a lit f.

Consequence (l, p) means that the lits in l imply the propagated lit f. The proof should be a proof of the clause "l implies f".

invariant: in Consequence (fun () -> l,p), all elements of l must be true in the current trail.

note on lazyiness: the justification is suspended (using unit -> …) to avoid potentially costly computations that might never be used if this literal is backtracked without participating in a conflict. Therefore the function that produces (l,p) needs only be safe in trails (partial models) that are conservative extensions of the current trail. If the theory isn't robust w.r.t. extensions of the trail (e.g. if its internal state undergoes significant changes), it can be easier to produce the explanation eagerly when propagating, and then use Consequence (fun () -> expl, proof) with the already produced (expl,proof) tuple.

type lbool =

| L_true

| L_false

| L_undefined
Valuation of an atom

module Clause_pool_id : sig ... end

module type ACTS = sig ... end

Actions available to the Plugin

type ('lit, 'proof) acts = (module ACTS with type lit = 'lit and type proof = 'proof)

The type for a slice of assertions to assume/propagate in the theory.

exception No_proof

module type LIT = sig ... end

module type PROOF = Sidekick_core.SAT_PROOF

module type PLUGIN_CDCL_T = sig ... end

Signature for theories to be given to the CDCL(T) solver

module type PLUGIN_SAT = sig ... end

Signature for pure SAT solvers

module type S = sig ... end

The external interface implemented by safe solvers, such as the one created by the Solver.Make and Mcsolver.Make functors.