Module PLUGIN_SAT.Proof

type t = proof

The stored proof (possibly nil, possibly on disk, possibly in memory)

type lit = lit

A boolean literal for the proof trace

type dproof = t -> unit

A delayed proof, used to produce proofs on demand from theories.

val with_proof : t -> (t -> unit) -> unit

If proof is enabled, call f on it to emit steps. if proof is disabled, the callback won't even be called.

val emit_input_clause : lit Iter.t -> t -> unit

Emit an input clause.

val emit_redundant_clause : lit Iter.t -> t -> unit

Emit a clause deduced by the SAT solver, redundant wrt axioms. The clause must be RUP wrt previous clauses.

val del_clause : lit Iter.t -> t -> unit

Forget a clause. Only useful for performance considerations.