Parameter Make_cdcl_t.1-Th

type t

The plugin state itself

type lit

module Lit : Sidekick_sat.Solver_intf.LIT with type t = lit

type proof

Proof storage/recording

module Proof : Sidekick_sat.Solver_intf.PROOF with type t = proof and type lit = lit

val push_level : t -> unit

Create a new backtrack level

val pop_levels : t -> int -> unit

Pop n levels of the theory

val partial_check : t -> (lit, proof) Sidekick_sat.Solver_intf.acts -> unit

Assume the lits in the slice, possibly using the slice to push new lits to be propagated or to raising a conflict or to add new lemmas.

val final_check : t -> (lit, proof) Sidekick_sat.Solver_intf.acts -> unit

Called at the end of the search in case a model has been found. If no new clause is pushed, then proof search ends and "sat" is returned; if lemmas are added, search is resumed; if a conflict clause is added, search backtracks and then resumes.