Module type Sidekick_sat__Solver_intf.SAT_STATE

type formula

val eval : formula -> bool

Returns the valuation of a formula in the current state of the sat solver.

raises UndecidedLit

if the literal is not decided

val eval_level : formula -> bool * int

Return the current assignement of the literals, as well as its decision level. If the level is 0, then it is necessary for the atom to have this value; otherwise it is due to choices that can potentially be backtracked.

raises UndecidedLit

if the literal is not decided

val iter_trail : (formula -> unit) -> unit

Iter through the formulas in order of decision/propagation (starting from the first propagation, to the last propagation).