Module Sidekick_smt_solver.Model_builder

Model Builder.

This contains a partial model, in construction. It is accessible to every theory, so they can contribute partial values.

TODO: seen values?

type t

include Sidekick_sigs.PRINT with type t := t

val pp : t Sidekick_sigs.printer

val create : Sidekick_core.Term.store -> t

val mem : t -> Sidekick_core.Term.t -> bool

val require_eval : t -> Sidekick_core.Term.t -> unit

Require that this term gets a value, and assign it to all terms in the given class.

val add : t -> ?subs:Sidekick_core.Term.t list -> Sidekick_core.Term.t -> Sigs.value -> unit

Add a value to the model.

parameter subs
if provided, these terms will be passed to require_eval to ensure they map to a value.

val gensym : t -> pre:string -> ty:Sidekick_core.Term.t -> Sidekick_core.Term.t

New fresh constant

type eval_cache = Sidekick_smt_solver__.Sigs.Term.Internal_.cache

val create_cache : int -> eval_cache

val eval : ?cache:eval_cache -> t -> Sidekick_core.Term.t -> Sigs.value

val eval_opt : ?cache:eval_cache -> t -> Sidekick_core.Term.t -> Sigs.value option

val pop_required : t -> Sidekick_core.Term.t option

gives the next subterm that is required but has no value yet

val to_map : ?cache:eval_cache -> t -> Sigs.value Sidekick_smt_solver__.Sigs.Term.Map.t