Module Sidekick_smt_solver.Preprocess

Preprocessor

The preprocessor turn mixed, raw literals (possibly simplified) into literals suitable for reasoning. Every literal undergoes preprocessing. Typically some clauses are also added to the solver on the side, and some subterms are found to be foreign variables.

type t

Preprocessor

val create : ?stat:Sidekick_util.Stat.t -> proof:Sidekick_proof.Tracer.t -> cc:Sidekick_cc.CC.t -> simplify:Sidekick_simplify.t -> Sidekick_core.Term.store -> t

module type PREPROCESS_ACTS = sig ... end

Actions given to preprocessor hooks

type preprocess_actions = (module PREPROCESS_ACTS)

Actions available to the preprocessor

type preprocess_hook = t -> is_sub:bool -> recurse:(Sigs.term -> Sigs.term) -> preprocess_actions -> Sigs.term -> Sigs.term option

Given a term, preprocess it.

The idea is to add literals and clauses to help define the meaning of the term, if needed. For example for boolean formulas, clauses for their Tseitin encoding can be added, with the formula acting as its own proxy symbol; or a new symbol might be added.

parameter preprocess_actions
actions available during preprocessing.

val on_preprocess : t -> preprocess_hook -> unit

Add a hook that will be called when terms are preprocessed

val simplify_and_preproc_lit : t -> preprocess_actions -> Sigs.lit -> Sigs.lit * Sigs.step_id option

val preprocess_clause : t -> preprocess_actions -> Sigs.lit list -> Sigs.step_id -> Sigs.lit list * Sigs.step_id

val preprocess_clause_array : t -> preprocess_actions -> Sigs.lit array -> Sigs.step_id -> Sigs.lit array * Sigs.step_id

val cc : t -> Sidekick_cc.CC.t