Module Sidekick_smt_solver.Sigs

Signature for the main SMT solver types.

Theories and concrete solvers rely on an environment that defines several important types:

sorts

terms (to represent logic expressions and formulas)

a congruence closure instance

a bridge to some SAT solver

In this module we collect signatures defined elsewhere and define the module types for the main SMT solver.

include module type of struct include Sidekick_core end

module Fmt = Sidekick_core.Fmt

Re-exports from core-logic

module Const = Sidekick_core.Const

module Term = Sidekick_core.Term

module Gensym = Sidekick_core.Gensym

view

module Bool_view = Sidekick_core.Bool_view

module CC_view = Sidekick_core.CC_view

module Default_cc_view = Sidekick_core.Default_cc_view

Main modules

module Bvar = Sidekick_core.Bvar

module Lit = Sidekick_core.Lit

module Proof_step = Sidekick_core.Proof_step

module Proof_core = Sidekick_core.Proof_core

module Proof_sat = Sidekick_core.Proof_sat

module Proof_trace = Sidekick_core.Proof_trace

module Proof_term = Sidekick_core.Proof_term

module Subst = Sidekick_core.Subst

module Var = Sidekick_core.Var

exception Resource_exhausted

module Simplify = Sidekick_simplify

module CC = Sidekick_cc.CC

module E_node = Sidekick_cc.E_node

module CC_expl = Sidekick_cc.Expl

type term = Term.t

type ty = term

type value = Term.t

type lit = Lit.t

type term_store = Term.store

type proof_trace = Proof_trace.t

type step_id = Proof_step.id

type sat_acts = Sidekick_sat.acts

type th_combination_conflict = {

lits : lit list;

semantic : (bool * term * term) list;

}

Conflict obtained during theory combination. It involves equalities merged because of the current model so it's not a "true" conflict and doesn't need to kill the current trail.

module type ARG = sig ... end

Argument to pass to the functor Make in order to create a new Msat-based SMT solver.