Module type Sigs.ARG

Argument to pass to the functor Make in order to create a new Msat-based SMT solver.

val view_as_cc : Sidekick_cc.view_as_cc

val is_valid_literal : Term.t -> bool

Is this a valid boolean literal? (e.g. is it a closed term, not inside a quantifier)