Module S.Model

Models

A model can be produced when the solver is found to be in a satisfiable state after a call to solve.

type t

val empty : t

val mem : t -> term -> bool

val find : t -> term -> term option

val eval : t -> term -> term option

val pp : t Sidekick_core.Fmt.printer