Module S.A

module S : Sidekick_core.SOLVER

type term = S.T.Term.t

val view_as_bool : term -> (term, term Iter.t) bool_view

Project the term into the boolean view.

val mk_bool : S.T.Term.store -> (term, term Sidekick_util.IArray.t) bool_view -> term

Make a term from the given boolean view.

val check_congruence_classes : bool

Configuration: add final-check handler to verify if new boolean formulas are present in the congruence closure. Only enable if some theories are susceptible to create boolean formulas during the proof search.

val lemma_bool_tauto : S.Lit.t Iter.t -> S.P.t -> unit

Boolean tautology lemma (clause)

val lemma_bool_c : string -> term list -> S.P.t -> unit

Basic boolean logic lemma for a clause |- c. proof_bool_c b name cs is the rule designated by name.

val lemma_bool_equiv : term -> term -> S.P.t -> unit

Boolean tautology lemma (equivalence)

val lemma_ite_true : a:term -> ite:term -> S.P.t -> unit

lemma a => ite a b c = b

val lemma_ite_false : a:term -> ite:term -> S.P.t -> unit

lemma ¬a => ite a b c = c

module Gensym : sig ... end

Fresh symbol generator.