Module Sidekick_th_data

Theory for datatypes.

type ('c, 't) data_view =

| T_cstor of 'c * 't Sidekick_util.IArray.t
T_cstor (c,args) is the term c(args)

| T_select of 'c * int * 't
T_select (c,i,u) means the i-th argument of u, which should start with constructor c

| T_is_a of 'c * 't
T_is_a (c,u) means u starts with constructor c

| T_other of 't
non-datatype term

Datatype-oriented view of terms.

'c is the representation of constructors

't is the representation of terms

type ('c, 'ty) data_ty_view =

| Ty_arrow of 'ty Iter.t * 'ty

| Ty_app of {

args : 'ty Iter.t;

}

| Ty_data of {

cstors : 'c;

}

| Ty_other

View of types in a way that is directly useful for the theory of datatypes

module type ARG = sig ... end

Argument to the functor

module type S = sig ... end

module Make : functor (A : ARG) -> S with module A = A