Module Sidekick_trace.Entry_id

Entry in the sink.

This integer tag represent a single entry in a trace, for example a line if we serialized using line-separate json values. In general each entry has its own unique ID that is monotonically increasing with time.

include sig ... end

type t = int

val equal : t -> t -> bool

val compare : t -> t -> int

val hash : int CCHash.t

val pp : int CCFormat.printer

val to_int : 'a -> 'a

val of_int_unsafe : int -> t

val dummy : t

Dummy id