Module Sidekick_util.Ser_value

Serialization representation.

A Ser_value.t describes how to serialized some structured data into bytes. It reflects the shape of the structured data but does not commit to a particular serialization format.

type t = private

| Null

| Bool of bool

| Str of string

| Bytes of string

| Int of int

| List of t list

| Dict of t Sidekick_util__.Util.Str_map.t

val null : t

val bool : bool -> t

val int : int -> t

val string : string -> t

val bytes : string -> t

val list : t list -> t

val dict : t Sidekick_util__.Util.Str_map.t -> t

val dict_of_list : (string * t) list -> t

val is_null : t -> bool

include Sidekick_sigs.PRINT with type t := t

val pp : t Sidekick_sigs.printer