Module Sidekick_core

Main Signatures.

Theories and concrete solvers rely on an environment that defines several important types:

types

terms (to represent logic expressions and formulas)

a congruence closure instance

a bridge to some SAT solver

In this module we define most of the main signatures used throughout Sidekick.

module Fmt = CCFormat

Re-exports from core-logic

module Const = Sidekick_core_logic.Const

module Str_const = Sidekick_core_logic.Str_const

module Term : sig ... end

view

module Bool_view : sig ... end

Boolean-oriented view of terms

module CC_view : sig ... end

View terms through the lens of a Congruence Closure

module Default_cc_view : sig ... end

Main modules

module Bvar = Sidekick_core_logic.Bvar

module Lit : sig ... end

Literals

module Subst = Sidekick_core_logic.Subst

module Var = Sidekick_core_logic.Var

module Box : sig ... end

module Gensym : sig ... end

Fresh symbol generation

exception Resource_exhausted

Const decoders for traces

val const_decoders : (string * Sidekick_core_logic.Const.Ops.t * (Sidekick_core_logic__.Types_.term Sidekick_util.Ser_decode.t -> Sidekick_core_logic.Const.view Sidekick_util.Ser_decode.t)) list