Module Sidekick_mini_cc.Make

Parameters

A : ARG

Signature

type term = A.T.Term.t

type fun_ = A.T.Fun.t

type term_state = A.T.Term.state

type t

val create : term_state -> t

val clear : t -> unit

Fully reset the congruence closure's state

val add_lit : t -> term -> bool -> unit

add_lit cc p sign asserts that p=sign

val check_sat : t -> bool

check_sat cc returns true if the current state is satisfiable, false if it's unsatisfiable

val classes : t -> term Iter.t Iter.t

Traverse the set of classes in the congruence closure. This should be called only if check returned Sat.