Module Sidekick_sat.Solver

The external interface implemented by SAT solvers.

type clause

type plugin = (module Sidekick_sat__Sigs.PLUGIN)

type solver

The main solver type.

type store

Stores atoms, clauses, etc.

module Clause : sig ... end

Main Solver Type

type t = solver

Main solver type, containing all state for solving.

val store : t -> store

Store for the solver

val stat : t -> Sidekick_util.Stat.t

Statistics

val proof : t -> Sidekick_core.Proof_trace.t

Access the inner proof

val on_conflict : t -> ( Clause.t, unit ) Sidekick_util.Event.t

val on_decision : t -> ( Sidekick_core.Lit.t, unit ) Sidekick_util.Event.t

val on_learnt : t -> ( Clause.t, unit ) Sidekick_util.Event.t

val on_gc : t -> ( Sidekick_core.Lit.t array, unit ) Sidekick_util.Event.t

Types

type res =

| Sat of (module Sidekick_sat__Sigs.SAT_STATE)	(*
Returned when the solver reaches SAT, with a model
*)

| Unsat of (module Sidekick_sat__Sigs.UNSAT_STATE with type clause = clause)	(*
Returned when the solver reaches UNSAT, with a proof
*)

Result type for the solver

exception UndecidedLit

Exception raised by the evaluating functions when a literal has not yet been assigned a value.

Base operations

val assume : t -> Sidekick_core.Lit.t list list -> unit

Add the list of clauses to the current set of assumptions. Modifies the sat solver state in place.

val add_clause : t -> Sidekick_core.Lit.t list -> Sidekick_core.Proof_step.id -> unit

Lower level addition of clauses

val add_clause_a : t -> Sidekick_core.Lit.t array -> Sidekick_core.Proof_step.id -> unit

Lower level addition of clauses

val add_input_clause : t -> Sidekick_core.Lit.t list -> unit

Like add_clause but with the justification of being an input clause

val add_input_clause_a : t -> Sidekick_core.Lit.t array -> unit

Like add_clause_a but with justification of being an input clause

Solving

val solve : ?on_progress:( unit -> unit ) -> ?assumptions:Sidekick_core.Lit.t list -> t -> res

Try and solves the current set of clauses.

parameter assumptions
additional atomic assumptions to be temporarily added. The assumptions are just used for this call to solve, they are not saved in the solver's state.

parameter on_progress
regularly called during solving. Can raise Resource_exhausted to stop solving.

raises Resource_exhausted
if the on_progress handler raised it to stop

Evaluating and adding literals

val add_lit : t -> ?default_pol:bool -> Sidekick_core.Lit.t -> unit

Ensure the SAT solver handles this particular literal, ie add a boolean variable for it if it's not already there.

val set_default_pol : t -> Sidekick_core.Lit.t -> bool -> unit

Set default polarity for the given boolean variable. Sign of the literal is ignored.

val true_at_level0 : t -> Sidekick_core.Lit.t -> bool

true_at_level0 a returns true if a was proved at level0, i.e. it must hold in all models

val eval_lit : t -> Sidekick_core.Lit.t -> Sidekick_sat__Sigs.lbool

Evaluate atom in current state

Assumption stack

val push_assumption : t -> Sidekick_core.Lit.t -> unit

Pushes an assumption onto the assumption stack. It will remain there until it's pop'd by pop_assumptions.

val pop_assumptions : t -> int -> unit

pop_assumptions solver n removes n assumptions from the stack. It removes the assumptions that were the most recently added via push_assumptions.

type propagation_result =

| PR_sat

| PR_conflict of {

backtracked : int;

}

| PR_unsat of (module Sidekick_sat__Sigs.UNSAT_STATE with type clause = clause)

Result returned by check_sat_propagations_only

val check_sat_propagations_only : ?assumptions:Sidekick_core.Lit.t list -> t -> propagation_result

check_sat_propagations_only solver uses the added clauses and local assumptions (using push_assumptions and assumptions) to quickly assess whether the context is satisfiable. It is not complete; calling solve is required to get an accurate result.

returns
either Ok() if propagation yielded no conflict, or Error c if a conflict clause c was found.

Initialization

val plugin_cdcl_t : (module Sidekick_sat__Sigs.THEORY_CDCL_T) -> (module Sidekick_sat__Sigs.PLUGIN)

val mk_plugin_cdcl_t : push_level:( unit -> unit ) -> pop_levels:( int -> unit ) -> ?partial_check:( (module Sidekick_sat__Sigs.ACTS) -> unit ) -> final_check:( (module Sidekick_sat__Sigs.ACTS) -> unit ) -> unit -> (module Sidekick_sat__Sigs.PLUGIN)

Create a plugin

parameter push_level
create a new backtrack level

parameter pop_levels
Pop n levels of the plugin

parameter partial_check
Assume the lits in the slice, possibly using the slice to push new lits to be propagated or to raising a conflict or to add new lemmas.

parameter final_check
Called at the end of the search in case a model has been found. If no new clause is pushed, then proof search ends and "sat" is returned; if lemmas are added, search is resumed; if a conflict clause is added, search backtracks and then resumes.

val create : ?stat:Sidekick_util.Stat.t -> ?size:[ `Tiny | `Small | `Big ] -> proof:Sidekick_core.Proof_trace.t -> plugin -> t

Create new solver

parameter theory
the theory

parameter the
proof

parameter size
the initial size of internal data structures. The bigger, the faster, but also the more RAM it uses.

val plugin_pure_sat : plugin

val create_pure_sat : ?stat:Sidekick_util.Stat.t -> ?size:[ `Tiny | `Small | `Big ] -> proof:Sidekick_core.Proof_trace.t -> unit -> t