Module Sidekick_smt_solver.Sigs

Signature for the main SMT solver types.

Theories and concrete solvers rely on an environment that defines several important types:

sorts

terms (to represent logic expressions and formulas)

a congruence closure instance

a bridge to some SAT solver

In this module we collect signatures defined elsewhere and define the module types for the main SMT solver.

include module type of struct include Sidekick_core end

module Fmt = Sidekick_core.Fmt

Re-exports from core-logic

module Const = Sidekick_core.Const

module Str_const = Sidekick_core.Str_const

module Term = Sidekick_core.Term

view

module Bool_view = Sidekick_core.Bool_view

module CC_view = Sidekick_core.CC_view

module Default_cc_view = Sidekick_core.Default_cc_view

Main modules

module Bvar = Sidekick_core.Bvar

module Lit = Sidekick_core.Lit

module Subst = Sidekick_core.Subst

module Var = Sidekick_core.Var

module Box = Sidekick_core.Box

module Gensym = Sidekick_core.Gensym

exception Resource_exhausted

Const decoders for traces

val const_decoders : (string * Sidekick_core_logic.Const.Ops.t * (Sidekick_core_logic__.Types_.term Sidekick_util.Ser_decode.t -> Sidekick_core_logic.Const.view Sidekick_util.Ser_decode.t)) list

module Simplify = Sidekick_simplify

module CC = Sidekick_cc.CC

module E_node = Sidekick_cc.E_node

module CC_expl = Sidekick_cc.Expl

module Proof = Sidekick_proof

type term = Term.t

type ty = term

type value = Term.t

type lit = Lit.t

type term_store = Term.store

type step_id = Sidekick_proof.Step.id

type sat_acts = Sidekick_sat.acts

type th_combination_conflict = {

lits : lit list;

semantic : (bool * term * term) list;

}

Conflict obtained during theory combination. It involves equalities merged because of the current model so it's not a "true" conflict and doesn't need to kill the current trail.

module type ARG = sig ... end

Argument to pass to the functor Make in order to create a new Msat-based SMT solver.