Module Sidekick_smt_solver.Theory

Signatures for theory plugins

module type S = sig ... end

A theory

type t = (module S)

A theory that can be used for this particular solver.

type 'a p = (module S with type t = 'a)

A theory that can be used for this particular solver, with state of type 'a.

val name : t -> string

Name of the theory

val make : name:string -> create_and_setup:(id:Theory_id.t -> Solver_internal.t -> 'st) -> ?push_level:('st -> unit) -> ?pop_levels:('st -> int -> unit) -> unit -> t